wip: move http method type to root
