+++ src/index.ts
@@ -13,7 +13,7 @@ export * from '@/Kernel';
 
 // Routing
 export * from '@/routing/deprecated-decorator/route';
-export type { HttpMethod } from '@/routing/declaration/http-method.type';
+export type { HttpMethod } from '@/routing/http-method.type';
 export type { RouteOptions } from '@/routing/declaration/route-options.type';
 export type { RouterMethod } from '@/routing/declaration/router-method.type';
 /**
