+++ src/routing/normalization/normalize-route-props.ts
@@ -1,9 +1,9 @@
 import type { HttpRequest, HttpRequestBase } from '@/Http';
-import type { HttpMethod } from '@/routing/declaration/http-method.type';
 import type { NormalizedRouteProps } from '@/routing/declaration/normalized-route-props.type';
 import type { RouteOptions } from '@/routing/declaration/route-options.type';
 import { RouteProps } from '@/routing/declaration/route-props.type';
 import type { RouterMethod } from '@/routing/declaration/router-method.type';
+import type { HttpMethod } from '@/routing/http-method.type';
 
 function resolveRouteOptions(options: RouteOptions): Pick<NormalizedRouteProps, 'parseBody' | 'bodyOptions'> {
   return {
